% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myc_matmult}
\alias{myc_matmult}
\title{myc_matmult}
\usage{
myc_matmult(A, B)
}
\arguments{
\item{A}{NumericMatrix}

\item{B}{NumericMatrix}
}
\value{
NumericMatrix
}
\description{
Matrix Multiplication
}
